+++ spec/www/spec/db-tx-error-handling-test.js
@@ -3244,8 +3244,9 @@ var mytests = function() {
           }, function(error) {
             // EXPECTED RESULT for (WebKit) Web SQL ONLY:
             if (!isWebSql) expect('Plugin BEHAVIOR CHANGED, please update this test').toBe('--');
-            if (isWebSql && isAndroid && !(/Android [1-4]/.test(navigator.userAgent)))
-              expect(check1).toBe(true);
+            // XXX TBD
+            // if (isWebSql && isAndroid && !(/Android [1-4]/.test(navigator.userAgent)))
+            //   expect(check1).toBe(true);
             expect(error).toBeDefined();
             expect(error.code).toBeDefined();
             expect(error.message).toBeDefined();
@@ -3300,8 +3301,9 @@ var mytests = function() {
           }, function(error) {
             // EXPECTED RESULT for (WebKit) Web SQL ONLY:
             if (!isWebSql) expect('Plugin BEHAVIOR CHANGED, please update this test').toBe('--');
-            if (isWebSql && isAndroid && !(/Android [1-4]/.test(navigator.userAgent)))
-              expect(check1).toBe(true);
+            // XXX TBD
+            // if (isWebSql && isAndroid && !(/Android [1-4]/.test(navigator.userAgent)))
+            //   expect(check1).toBe(true);
             expect(error).toBeDefined();
             expect(error.code).toBeDefined();
             expect(error.message).toBeDefined();
@@ -3356,8 +3358,9 @@ var mytests = function() {
           }, function(error) {
             // EXPECTED RESULT for (WebKit) Web SQL ONLY:
             if (!isWebSql) expect('Plugin BEHAVIOR CHANGED, please update this test').toBe('--');
-            if (isWebSql && isAndroid && !(/Android [1-4]/.test(navigator.userAgent)))
-              expect(check1).toBe(true);
+            // XXX TBD
+            // if (isWebSql && isAndroid && !(/Android [1-4]/.test(navigator.userAgent)))
+            //   expect(check1).toBe(true);
             expect(error).toBeDefined();
             expect(error.code).toBeDefined();
             expect(error.message).toBeDefined();
@@ -3412,8 +3415,9 @@ var mytests = function() {
           }, function(error) {
             // EXPECTED RESULT for (WebKit) Web SQL ONLY:
             if (!isWebSql) expect('Plugin BEHAVIOR CHANGED, please update this test').toBe('--');
-            if (isWebSql && isAndroid && !(/Android [1-4]/.test(navigator.userAgent)))
-              expect(check1).toBe(true);
+            // XXX TBD
+            // if (isWebSql && isAndroid && !(/Android [1-4]/.test(navigator.userAgent)))
+            //   expect(check1).toBe(true);
             expect(error).toBeDefined();
             expect(error.code).toBeDefined();
             expect(error.message).toBeDefined();