+++ spec/www/spec/db-tx-sql-features-test.js
@@ -49,7 +49,7 @@ var mytests = function() {
 
         // Known to work with:
         // - iOS 9 Web SQL
-        // - Android (default Android-sqlite-connector implementation)
+        // - Android (default Android evcore-native-driver implementation)
         // - iOS & Windows (with newer sqlite3 build)
         it(suiteName + 'db readTransaction with a WITH clause', function(done) {
           if (isWP8) pending('NOT IMPLEMENTED for WP(8)');
@@ -169,7 +169,7 @@ var mytests = function() {
 
         it(suiteName + 'Basic JSON1 json test', function(done) {
           if (isWebSql) pending('SKIP for Web SQL (not implemented)');
-          if (!isWebSql && isAndroid && isImpl2) pending('SKIP for androidDatabaseImplementation: 2 [NOT IMPLEMENTED]');
+          if (!isWebSql && !isWindows && isAndroid && isImpl2) pending('TBD SKIP for androidDatabaseImplementation: 2'); // XXX TBD
 
           var db = openDatabase('basic-json1-json-test.db', '1.0', 'Test', DEFAULT_SIZE);
 
@@ -198,7 +198,7 @@ var mytests = function() {
 
         it(suiteName + 'JSON1 json_object test', function(done) {
           if (isWebSql) pending('SKIP for Web SQL (not implemented)');
-          if (!isWebSql && isAndroid && isImpl2) pending('SKIP for androidDatabaseImplementation: 2 [NOT IMPLEMENTED]');
+          if (!isWebSql && !isWindows && isAndroid && isImpl2) pending('TBD SKIP for androidDatabaseImplementation: 2'); // XXX TBD
 
           var db = openDatabase('json1-json-object-test.db', '1.0', 'Test', DEFAULT_SIZE);
 
@@ -228,7 +228,7 @@ var mytests = function() {
 
         it(suiteName + 'create virtual table using FTS5', function(done) {
           if (isWebSql) pending('SKIP for Web SQL (not implemented)');
-          if (!isWebSql && isAndroid && isImpl2) pending('SKIP for androidDatabaseImplementation: 2 [NOT IMPLEMENTED]');
+          if (!isWebSql && !isWindows && isAndroid && isImpl2) pending('TBD SKIP for androidDatabaseImplementation: 2'); // XXX TBD
 
           var db = openDatabase('virtual-table-using-fts5.db', '1.0', 'Test', DEFAULT_SIZE);